Split up redux files
Split files in `client/src/redux` (-> `client/src/state`)
into smaller files.
