+++ client/src/containers/app/AppHelper.js
@@ -14,13 +14,21 @@ limitations under the License.
 ----------------------------------------------------------------------------- */
 
 
-import AppActions, { AppSelectors } from 'redux/AppRedux';
-import ApproverActions, { ApproverSelectors } from 'redux/ApproverRedux';
-import AuthActions, { AuthSelectors } from 'redux/AuthRedux';
-import ChatActions, { ChatSelectors } from 'redux/ChatRedux';
-import UserActions, { UserSelectors } from 'redux/UserRedux';
-import RequesterActions, {
-  RequesterSelectors } from 'redux/RequesterRedux';
+import {
+  AppActions,
+  ApproverActions,
+  AuthActions,
+  ChatActions,
+  RequesterActions,
+  UserActions,
+
+  AppSelectors,
+  ApproverSelectors,
+  AuthSelectors,
+  ChatSelectors,
+  RequesterSelectors,
+  UserSelectors,
+} from 'state';
 
 
 //
