+++ client/src/containers/signup/Signup.js
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 
 
 import './Signup.css';
-import AuthActions, { AuthSelectors } from 'redux/AuthRedux';
+import { AuthActions, AuthSelectors } from 'state';
 import SignupForm from 'components/forms/SignupForm';
 import logo from 'images/next-logo-billboard.png';
