--- client/src/redux/index.js
@@ -1,39 +1,0 @@
-/* Copyright 2018 Contributors to Hyperledger Sawtooth
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
------------------------------------------------------------------------------ */
-
-
-import { combineReducers } from 'redux';
-import { loadingBarReducer } from 'react-redux-loading-bar';
-
-
-import { reducer as AppReducer } from './AppRedux';
-import { reducer as ApproverReducer } from './ApproverRedux';
-import { reducer as AuthReducer } from './AuthRedux';
-import { reducer as ChatReducer } from './ChatRedux';
-import { reducer as RequesterReducer } from './RequesterRedux';
-import { reducer as UserReducer } from './UserRedux';
-
-
-const reducers = combineReducers({
-  app:        AppReducer,
-  approver:   ApproverReducer,
-  auth:       AuthReducer,
-  chat:       ChatReducer,
-  requester:  RequesterReducer,
-  user:       UserReducer,
-  loadingBar: loadingBarReducer,
-});
-
-export default reducers;
