+++ client/src/sagas/index.js
@@ -21,14 +21,13 @@ import API from 'services/Api';
 import FixtureAPI from 'services/FixtureApi';
 
 
-import { AppTypes } from 'redux/AppRedux';
-import { ApproverTypes } from 'redux/ApproverRedux';
-import { AuthTypes } from 'redux/AuthRedux';
-import { ChatTypes } from 'redux/ChatRedux';
-import { RequesterTypes } from 'redux/RequesterRedux';
-import { UserTypes } from 'redux/UserRedux';
-
-
+import {
+  AppTypes,
+  ApproverTypes,
+  AuthTypes,
+  ChatTypes,
+  RequesterTypes,
+  UserTypes } from 'state';
 import {
   approveProposals,
   rejectProposals,
