+++ client/src/state/requester/RequesterSelectors.js
@@ -14,87 +14,9 @@ limitations under the License.
 ----------------------------------------------------------------------------- */
 
 
-import { createReducer, createActions } from 'reduxsauce';
-import Immutable from 'seamless-immutable';
 import * as utils from 'services/Utils';
 
 
-//
-// Actions
-//
-//
-//
-//
-const { Types, Creators } = createActions({
-  allPacksRequest:      ['start', 'limit'],
-  allPacksSuccess:      ['packs', 'packsTotalCount'],
-  allPacksFailure:      ['error'],
-
-  allRolesRequest:      ['start', 'limit'],
-  allRolesSuccess:      ['roles', 'rolesTotalCount'],
-  allRolesFailure:      ['error'],
-
-  roleRequest:          ['id'],
-  rolesRequest:         ['ids'],
-  roleSuccess:          ['role'],
-  roleFailure:          ['error'],
-
-  baseRequest:          null,
-  baseSuccess:          ['base'],
-  baseFailure:          ['error'],
-
-  packRequest:          ['id'],
-  packsRequest:         ['ids'],
-  packSuccess:          ['pack'],
-  packFailure:          ['error'],
-
-  proposalRequest:      ['id'],
-  proposalSuccess:      ['proposal'],
-  proposalFailure:      ['error'],
-  proposalsRequest:     ['ids'],
-
-  packAccessRequest:    ['id', 'userId', 'reason'],
-  packAccessSuccess:    null,
-  packAccessFailure:    null,
-
-  roleAccessRequest:    ['id', 'userId', 'reason'],
-  roleAccessSuccess:    null,
-  roleAccessFailure:    null,
-
-  resetAll:             null,
-});
-
-
-export const RequesterTypes = Types;
-export default Creators;
-
-//
-// State
-//
-//
-//
-//
-export const INITIAL_STATE = Immutable({
-  activeProposal:   null,
-  activeRole:       null,
-  error:            null,
-  fetching:         null,
-  fetchingAllRoles: null,
-  fetchingAllPacks: null,
-  packs:            null,
-  recommended:      null,
-  requests:         null,
-  roles:            null,
-  rolesTotalCount:  null,
-  packsTotalCount:  null,
-});
-
-//
-// Selectors
-//
-//
-//
-//
 export const RequesterSelectors = {
   rolesTotalCount: (state) => state.requester.rolesTotalCount,
   packsTotalCount: (state) => state.requester.packsTotalCount,
@@ -132,6 +54,8 @@ export const RequesterSelectors = {
         proposal.object_id === role.id
       )
     ).slice(0, 3),
+
+
   // Retrieve recommended packs
   recommendedPacks: (state) => {
     if (!state.requester.recommended || !state.user.me) return null;
@@ -283,6 +207,7 @@ export const RequesterSelectors = {
     return result;
   },
 
+
   /**
    * Find the proposal IDs of a pack
    * @param   {object} state      Redux state
@@ -301,126 +226,3 @@ export const RequesterSelectors = {
     ).map(item => item.proposal_id);
   },
 };
-
-//
-// Reducers
-// General
-//
-//
-//
-export const request = {
-  allRoles:   (state) => state.merge({ fetchingAllRoles: true }),
-  allPacks:   (state) => state.merge({ fetchingAllPacks: true }),
-  temp:       (state) => state.merge({ fetching: true }),
-};
-
-export const failure = (state, { error }) => {
-  return state.merge({ fetching: false, error });
-};
-export const resetAll = () => {
-  return INITIAL_STATE;
-};
-
-//
-// Reducers
-// Success
-//
-//
-//
-export const success = {
-  access: (state) =>
-    state.merge({
-      fetching: false,
-    }),
-  allRoles: (state, { roles, rolesTotalCount }) =>
-    state.merge({
-      fetchingAllRoles: false,
-      rolesTotalCount,
-      roles: utils.merge(
-        state.roles || [], roles || []
-      ),
-    }),
-  allPacks: (state, { packs, packsTotalCount }) =>
-    state.merge({
-      fetchingAllPacks: false,
-      packsTotalCount,
-      packs: utils.merge(
-        state.packs || [], packs || []
-      ),
-    }),
-  base: (state, { base }) =>
-    state.merge({
-      fetching: false,
-      recommended: utils.merge(
-        state.recommended || [], base[0].data.data || []
-      ),
-      roles: utils.merge(
-        state.roles || [], base[0].data.data || []
-      ),
-    }),
-  pack: (state, { pack }) =>
-    state.merge({
-      fetching: false,
-      packs: utils.merge(
-        state.packs || [], [pack]
-      ),
-    }),
-  proposal: (state, { proposal }) =>
-    state.merge({
-      fetching: false,
-      requests: utils.merge(
-        state.requests || [], [proposal]
-      ),
-    }),
-  role: (state, { role }) => {
-    return state.merge({
-      fetching: false,
-      roles: utils.merge(
-        state.roles || [], [role]
-      ),
-    });
-  },
-};
-
-//
-// Hooks
-//
-//
-//
-//
-export const reducer = createReducer(INITIAL_STATE, {
-  [Types.RESET_ALL]:              resetAll,
-  [Types.BASE_REQUEST]:           request.temp,
-  [Types.BASE_SUCCESS]:           success.base,
-  [Types.BASE_FAILURE]:           failure,
-
-  // Roles
-  [Types.ALL_ROLES_REQUEST]:      request.allRoles,
-  [Types.ALL_ROLES_SUCCESS]:      success.allRoles,
-  [Types.ALL_ROLES_FAILURE]:      failure,
-  [Types.ROLES_REQUEST]:          request.temp,
-  [Types.ROLE_REQUEST]:           request.temp,
-  [Types.ROLE_SUCCESS]:           success.role,
-  [Types.ROLE_FAILURE]:           failure,
-  [Types.ROLE_ACCESS_REQUEST]:    request.temp,
-  [Types.ROLE_ACCESS_SUCCESS]:    success.access,
-  [Types.ROLE_ACCESS_FAILURE]:    failure,
-
-  // Packs
-  [Types.ALL_PACKS_REQUEST]:      request.allPacks,
-  [Types.ALL_PACKS_SUCCESS]:      success.allPacks,
-  [Types.ALL_PACKS_FAILURE]:      failure,
-  [Types.PACKS_REQUEST]:          request.temp,
-  [Types.PACK_REQUEST]:           request.temp,
-  [Types.PACK_SUCCESS]:           success.pack,
-  [Types.PACK_FAILURE]:           failure,
-  [Types.PACK_ACCESS_REQUEST]:    request.temp,
-  [Types.PACK_ACCESS_SUCCESS]:    success.access,
-  [Types.PACK_ACCESS_FAILURE]:    failure,
-
-  // Proposals
-  [Types.PROPOSALS_REQUEST]:      request.temp,
-  [Types.PROPOSAL_REQUEST]:       request.temp,
-  [Types.PROPOSAL_SUCCESS]:       success.proposal,
-  [Types.PROPOSAL_FAILURE]:       failure,
-});
