+++ client/src/state/user/UserSelectors.js
@@ -14,21 +14,16 @@ limitations under the License.
 ----------------------------------------------------------------------------- */
 
 
-import { applyMiddleware, compose, createStore } from 'redux';
-import createSagaMiddleware from 'redux-saga';
+import * as storage from 'services/Storage';
 
 
-import sagas from './sagas';
-import reducers from './state';
-
-
-export let store;
-
-
-export const create = () => {
-  const sagaMiddleware = createSagaMiddleware();
-  store = createStore(reducers, compose(applyMiddleware(sagaMiddleware)));
-
-  sagaMiddleware.run(sagas);
-  return store;
+export const UserSelectors = {
+  me:         (state) => state.user.me,
+  id:         (state) =>
+    (state.user.me && state.user.me.id) || storage.getUserId(),
+  users:      (state) => state.user.users,
+  memberOf:   (state) => state.user.me && state.user.me.memberOf,
+  userFromId: (state, id) =>
+    state.user.users &&
+    state.user.users.find(user => user.id === id),
 };
